+++ examples/fuse_examples/imaging/classification/duke/runner_duke.py
@@ -17,357 +17,199 @@ Created on June 30, 2021
 
 """
 
-import logging
 import os
-from typing import OrderedDict, Optional, List, Dict, Any
+import copy
+from typing import Any, OrderedDict
+import logging
 
-import torch.nn.functional as F
+import torch
+import pytorch_lightning as pl
 import torch.optim as optim
 from torch.utils.data.dataloader import DataLoader
+import torch.nn.functional as F
 
+from fuse.utils.utils_debug import FuseDebug
 import fuse.utils.gpu as GPU
-import fuseimg.datasets.duke.duke_label_type as duke_label_type
-from fuse_examples.fuse_examples_utils import ask_user
-from fuse_examples.imaging.classification import duke
-
-from fuse_examples.imaging.utils.backbone_3d_multichannel import Fuse_model_3d_multichannel, ResNet
-from fuse.data.utils.collates import CollateDefault
-from fuse.data.utils.samplers import BatchSamplerDefault
+from fuse.utils.utils_logger import fuse_logger_start
+from fuse.utils.file_io.file_io import create_dir, save_dataframe, load_pickle
 from fuse.data.utils.split import dataset_balanced_division_to_folds
+
+# import fuse_examples.imaging.classification.duke.duke_utils as duke_utils
+from fuse_examples.imaging.utils.backbone_3d_multichannel import Fuse_model_3d_multichannel, ResNet
+from fuse.dl.models.heads import Head1DClassifier
 from fuse.dl.losses.loss_default import LossDefault
 
-# from fuse.dl.managers.callbacks.callback_metric_statistics import MetricStatisticsCallback
-# from fuse.dl.managers.callbacks.callback_tensorboard import TensorboardCallback
-# from fuse.dl.managers.callbacks.callback_time_statistics import TimeStatisticsCallback
-# from fuse.dl.managers.manager_default import ManagerDefault
-from fuse.dl.models.heads import Head1DClassifier
-from fuse.eval.evaluator import EvaluatorDefault
-from fuse.eval.metrics.classification.metrics_classification_common import MetricAccuracy, MetricAUCROC, MetricROCCurve
+from fuse.data.utils.collates import CollateDefault
+from fuse.data.utils.samplers import BatchSamplerDefault
 from fuse.eval.metrics.classification.metrics_thresholding_common import MetricApplyThresholds
-from fuse.utils.file_io.file_io import load_pickle
-from fuse.utils.rand.seed import Seed
-from fuse.utils.ndict import NDict
-from fuse.utils.utils_logger import fuse_logger_start
+from fuse.eval.metrics.classification.metrics_classification_common import MetricAccuracy, MetricAUCROC, MetricROCCurve
+
+from fuse.dl.lightning.pl_module import LightningModuleDefault
+from fuse.dl.lightning.pl_funcs import convert_predictions_to_dataframe
+
+from fuse.eval.evaluator import EvaluatorDefault
 from fuseimg.datasets.duke.duke import Duke
+from fuseimg.datasets.duke.duke_label_type import DukeLabelType
 
 
-def main() -> None:
-    mode = "default"  # Options: 'default', 'fast', 'debug', 'verbose', 'user'. See details in FuseDebug
+##########################################
+# Debug modes
+##########################################
+mode = "default"  # Options: 'default', 'debug'. See details in FuseDebug
+debug = FuseDebug(mode)
 
-    # allocate gpus
-    # To use cpu - set NUM_GPUS to 0
-    if mode == "debug":
-        NUM_GPUS = 1
-    else:
-        NUM_GPUS = 2
-    # uncomment if you want to use specific gpus instead of automatically looking for free ones
-    force_gpus = None  # [0]
-    GPU.choose_and_enable_multiple_gpus(NUM_GPUS, force_gpus=force_gpus)
+##########################################
+# Output Paths
+##########################################
+# ROOT = duke_utils.get_duke_user_dir()
+ROOT = "/tmp/_duke_sagi"
+data_dir = os.environ["DUKE_DATA_PATH"]
 
-    if False:
-        selected_positive = [1, 2, 3, 5, 6, 10, 12, 596, 900, 901]
-        selected_negative = [4, 6, 7, 8, 11, 13, 14, 120, 902, 903]
+# if mode == "debug": # super temp :)
+if True:  # super temp :)
+    data_split_file = "DUKE_folds_debug.pkl"
 
-        selected_sample_ids = [f"Breast_MRI_{ii:03d}" for ii in selected_positive + selected_negative]
-    else:
-        selected_sample_ids = None
+    selected_positive = [1, 2, 3, 5, 6, 10, 12, 596, 900, 901]
+    selected_negative = [4, 6, 7, 8, 11, 13, 14, 120, 902, 903]
+    selected_sample_ids = [f"Breast_MRI_{ii:03d}" for ii in selected_positive + selected_negative]
 
-    PATHS, TRAIN_COMMON_PARAMS, INFER_COMMON_PARAMS, EVAL_COMMON_PARAMS = get_setting(
-        mode, selected_sample_ids=selected_sample_ids
-    )
-    print(PATHS)
+    model_dir = os.path.join(ROOT, "model_dir_debug")
+    cache_dir = os.path.join(ROOT, "cache_dir_debug")
+    force_reset_model_dir = True
 
-    RUNNING_MODES = ["train", "infer", "eval"]  # Options: 'train', 'infer', 'eval'
-    # train
-    if "train" in RUNNING_MODES:
-        print(TRAIN_COMMON_PARAMS)
-        run_train(paths=PATHS, train_params=TRAIN_COMMON_PARAMS)
+    num_workers = 10
+    batch_size = 2
+    num_epochs = 2
+    num_devices = 1
 
-    # infer
-    if "infer" in RUNNING_MODES:
-        print(INFER_COMMON_PARAMS)
-        run_infer(paths=PATHS, infer_common_params=INFER_COMMON_PARAMS, audit_cache="train" not in RUNNING_MODES)
+else:
+    data_split_file = "DUKE_folds.pkl"
 
-    # eval
-    if "eval" in RUNNING_MODES:
-        print(EVAL_COMMON_PARAMS)
-        run_eval(paths=PATHS, eval_common_params=EVAL_COMMON_PARAMS)
+    selected_sample_ids = None
 
-    print(f"Done running with heldout={INFER_COMMON_PARAMS['data.infer_folds']}")
+PATHS = {
+    "model_dir": model_dir,
+    "force_reset_model_dir": force_reset_model_dir,  # If True will reset model dir automatically - otherwise will prompt 'are you sure' message.
+    "data_dir": data_dir,
+    "cache_dir": cache_dir,
+    "inference_dir": os.path.join(model_dir, "infer"),
+    "eval_dir": os.path.join(model_dir, "eval"),
+    "data_split_filename": os.path.join(ROOT, data_split_file),
+}
 
+##########################################
+# Train Common Params
+##########################################
+num_folds = 5
+heldout_fold = 4
+label_type: str = DukeLabelType.STAGING_TUMOR_SIZE
 
-def get_setting(
-    mode: str,
-    label_type: str = duke_label_type.DukeLabelType.STAGING_TUMOR_SIZE,
-    n_folds: int = 5,
-    heldout_fold: int = 4,
-    selected_sample_ids: List[str] = None,
-    num_epoch: Optional[int] = None,
-) -> List[Dict[str, Any]]:
-    ###########################################################################################################
-    # Fuse
-    ###########################################################################################################
-    ##########################################
-    # Debug modes
-    ##########################################
-
-    # debug = FuseDebug(mode)
-    ##########################################
-    # Output Paths
-    ##########################################
-    assert (
-        "DUKE_DATA_PATH" in os.environ
-    ), "Expecting environment variable DUKE_DATA_PATH to be set. Follow the instruction in example README file to download and set the path to the data"
-    ROOT = duke.get_duke_user_dir()
-
-    data_dir = os.environ["DUKE_DATA_PATH"]
-
-    if mode == "debug":
-        data_split_file = os.path.join(ROOT, "DUKE_folds_debug.pkl")
-        if selected_sample_ids is None:
-            selected_sample_ids = duke.get_samples_for_debug(
-                data_dir=data_dir, n_pos=10, n_neg=10, label_type=label_type, sample_ids=duke.get_selected_sample_ids()
-            )
-        print(selected_sample_ids)
-        cache_dir = os.path.join(ROOT, "cache_dir_debug")
-        model_dir = os.path.join(ROOT, "model_dir_debug")
-        num_workers = 0
-        batch_size = 2
-        if num_epoch is None:
-            num_epoch = 5
-    else:
-        data_split_file = os.path.join(ROOT, "DUKE_folds_v5.pkl")
-        cache_dir = os.path.join(ROOT, "cache_dir_v5")
-        model_dir = os.path.join(ROOT, "model_dir_v5")
-
-        num_workers = 16  # put 0 to debug
-        batch_size = 50
-        if num_epoch is None:
-            num_epoch = 20  # 150
-
-    PATHS = {
-        "model_dir": model_dir,
-        "force_reset_model_dir": True,  # If True will reset model dir automatically - otherwise will prompt 'are you sure' message.
-        "cache_dir": cache_dir,
-        "data_split_filename": os.path.join(ROOT, data_split_file),
-        "data_dir": data_dir,
-        "inference_dir": os.path.join(model_dir, "infer_dir"),
-        "eval_dir": os.path.join(model_dir, "eval_dir"),
-    }
-
-    ##########################################
-    # Train Common Params
-    ##########################################
-    TRAIN_COMMON_PARAMS = {}
-    # ============
-    # Model
-    # ============
-
-    # ============
-    # Data
-    # ============
-
-    train_folds = [i % n_folds for i in range(heldout_fold + 1, heldout_fold + n_folds - 1)]
-    validation_fold = (heldout_fold - 1) % n_folds
-    TRAIN_COMMON_PARAMS["data.selected_sample_ids"] = selected_sample_ids
-    TRAIN_COMMON_PARAMS["data.batch_size"] = batch_size
-    TRAIN_COMMON_PARAMS["data.train_num_workers"] = num_workers
-    TRAIN_COMMON_PARAMS["data.validation_num_workers"] = num_workers
-    TRAIN_COMMON_PARAMS["data.num_folds"] = n_folds
-    TRAIN_COMMON_PARAMS["data.train_folds"] = train_folds
-    TRAIN_COMMON_PARAMS["data.validation_folds"] = [validation_fold]
-
-    # ===============
-    # Manager - Train
-    # ===============
-    TRAIN_COMMON_PARAMS["manager.train_params"] = {
-        "num_epochs": num_epoch,
-        "virtual_batch_size": 1,  # number of batches in one virtual batch
-        "start_saving_epochs": 10,  # first epoch to start saving checkpoints from
-        "gap_between_saving_epochs": 5,  # number of epochs between saved checkpoint
-    }
-    TRAIN_COMMON_PARAMS["manager.best_epoch_source"] = {
-        "source": "metrics.auc",  # can be any key from 'epoch_results'
-        "optimization": "max",  # can be either min/max
-        "on_equal_values": "better",
-        # can be either better/worse - whether to consider best epoch when values are equal
-    }
-    TRAIN_COMMON_PARAMS["manager.learning_rate"] = 1e-5
-    TRAIN_COMMON_PARAMS["manager.weight_decay"] = 0.001
-    TRAIN_COMMON_PARAMS["manager.dropout"] = 0.5
-    TRAIN_COMMON_PARAMS["manager.momentum"] = 0.9
-    TRAIN_COMMON_PARAMS["manager.resume_checkpoint_filename"] = None  # if not None, will try to load the checkpoint
-    TRAIN_COMMON_PARAMS["imaging_dropout"] = 0.25
-    # TRAIN_COMMON_PARAMS['fused_dropout'] = 0.0
-    # TRAIN_COMMON_PARAMS['clinical_dropout'] = 0.0
-
-    TRAIN_COMMON_PARAMS["num_backbone_features_imaging"] = 512
-
-    # in order to add relevant tabular feature uncomment:
-    # num_backbone_features_clinical, post_concat_inputs,post_concat_model
-    TRAIN_COMMON_PARAMS["num_backbone_features_clinical"] = None  # 256
-    TRAIN_COMMON_PARAMS["post_concat_inputs"] = None  # [('data.clinical_features',9),]
-    TRAIN_COMMON_PARAMS["post_concat_model"] = None  # (256,256)
-
-    if TRAIN_COMMON_PARAMS["num_backbone_features_clinical"] is None:
-        TRAIN_COMMON_PARAMS["num_backbone_features"] = TRAIN_COMMON_PARAMS["num_backbone_features_imaging"]
-    else:
-        TRAIN_COMMON_PARAMS["num_backbone_features"] = (
-            TRAIN_COMMON_PARAMS["num_backbone_features_imaging"] + TRAIN_COMMON_PARAMS["num_backbone_features_clinical"]
-        )
-
-    # classification_task:
-    # supported tasks are: 'Staging Tumor Size','Histology Type','is High Tumor Grade Total','PCR'
-    TRAIN_COMMON_PARAMS["classification_task"] = label_type
-    TRAIN_COMMON_PARAMS["class_num"] = label_type.get_num_classes()
-
-    # backbone parameters
-    TRAIN_COMMON_PARAMS["backbone_model_dict"] = {
-        "input_channels_num": 1,
-    }
-
-    ######################################
-    # Inference Common Params
-    ######################################
-    INFER_COMMON_PARAMS = {}
-    INFER_COMMON_PARAMS["infer_filename"] = "validation_set_infer.gz"
-    INFER_COMMON_PARAMS["checkpoint"] = "best"  # Fuse TIP: possible values are 'best', 'last' or epoch_index.
-    INFER_COMMON_PARAMS["data.infer_folds"] = [heldout_fold]  # infer validation set
-    INFER_COMMON_PARAMS["data.batch_size"] = 4
-    INFER_COMMON_PARAMS["data.num_workers"] = num_workers
-    INFER_COMMON_PARAMS["classification_task"] = TRAIN_COMMON_PARAMS["classification_task"]
-
-    ######################################
-    # Analyze Common Params
-    ######################################
-    EVAL_COMMON_PARAMS = {}
-    EVAL_COMMON_PARAMS["infer_filename"] = INFER_COMMON_PARAMS["infer_filename"]
-
-    return PATHS, TRAIN_COMMON_PARAMS, INFER_COMMON_PARAMS, EVAL_COMMON_PARAMS
+train_folds = [i % num_folds for i in range(heldout_fold + 1, heldout_fold + num_folds - 1)]
+validation_folds = [(heldout_fold - 1) % num_folds]
 
 
-#################################
-# Train Template
-#################################
-def run_train(
-    paths: dict, train_params: dict, reset_cache: Optional[bool] = None, audit_cache: Optional[bool] = None
-) -> None:
-    Seed.set_seed(222, False)
-    # ==============================================================================
-    # Logger
-    # ==============================================================================
-    fuse_logger_start(output_path=paths["model_dir"], console_verbose_level=logging.INFO)
-    lgr = logging.getLogger("Fuse")
-    lgr.info("Fuse Train", {"attrs": ["bold", "underline"]})
+TRAIN_COMMON_PARAMS = {}
+# ============
+# Data
+# ============
+TRAIN_COMMON_PARAMS["data.batch_size"] = batch_size
+TRAIN_COMMON_PARAMS["data.train_num_workers"] = num_workers
+TRAIN_COMMON_PARAMS["data.validation_num_workers"] = num_workers
+TRAIN_COMMON_PARAMS["data.cache_num_workers"] = num_workers
+TRAIN_COMMON_PARAMS["data.sample_ids"] = selected_sample_ids
+TRAIN_COMMON_PARAMS["data.num_folds"] = num_folds
+TRAIN_COMMON_PARAMS["data.train_folds"] = train_folds
+TRAIN_COMMON_PARAMS["data.validation_folds"] = validation_folds
 
-    lgr.info(f'model_dir={paths["model_dir"]}', {"color": "magenta"})
-    lgr.info(f'cache_dir={paths["cache_dir"]}', {"color": "magenta"})
-    lgr.info(f'train folds={train_params["data.train_folds"]}', {"color": "magenta"})
-    lgr.info(f'validation folds={train_params["data.validation_folds"]}', {"color": "magenta"})
+# ===============
+# PL Trainer
+# ===============
+TRAIN_COMMON_PARAMS["trainer.num_epochs"] = num_epochs
+TRAIN_COMMON_PARAMS["trainer.num_devices"] = num_devices
+TRAIN_COMMON_PARAMS["trainer.accelerator"] = "gpu"
+TRAIN_COMMON_PARAMS["trainer.ckpt_path"] = None
 
-    # ==============================================================================
-    # Data
-    # ==============================================================================
-    # Train Data
-    lgr.info("Train Data:", {"attrs": "bold"})
+# ===============
+# Optimizer
+# ===============
+TRAIN_COMMON_PARAMS["opt.lr"] = 1e-5
+TRAIN_COMMON_PARAMS["opt.weight_decay"] = 1e-3
 
-    if reset_cache is None:
-        reset_cache = ask_user("Do you want to reset cache?")
-    cache_kwargs = {"use_pipeline_hash": False}
-    if not reset_cache:
-        if audit_cache is None:
-            audit_cache = ask_user("Do you want to audit cache?")
-        if not audit_cache:
-            cache_kwargs2 = dict(audit_first_sample=False, audit_rate=None)
-            cache_kwargs = {**cache_kwargs, **cache_kwargs2}
+# ===================================================================================================================
+# Model
+# ===================================================================================================================
+TRAIN_COMMON_PARAMS["model.momentum"] = 0.9
+TRAIN_COMMON_PARAMS["model.dropout_rate"] = 0.5
+TRAIN_COMMON_PARAMS["model.imaging_dropout"] = 0.25
+TRAIN_COMMON_PARAMS["model.post_concat_inputs"] = None  # [('data.clinical_features',9),]
+TRAIN_COMMON_PARAMS["model.post_concat_model"] = None  # (256,256)
 
-    # split to folds randomly
-    params = dict(
-        label_type=train_params["classification_task"],
-        data_dir=paths["data_dir"],
-        cache_dir=paths["cache_dir"],
-        reset_cache=reset_cache,
-        sample_ids=train_params["data.selected_sample_ids"],
-        num_workers=train_params["data.train_num_workers"],
-        cache_kwargs=cache_kwargs,
-        train=False,
-        verbose=False,
+TRAIN_COMMON_PARAMS["model.classification_task"] = label_type
+TRAIN_COMMON_PARAMS["model.class_num"] = label_type.get_num_classes()
+## Backbone parameters
+TRAIN_COMMON_PARAMS["model.bb.input_channels_num"] = 1
+TRAIN_COMMON_PARAMS["model.bb.num_features_imaging"] = 512
+TRAIN_COMMON_PARAMS["model.bb.num_features_clinical"] = None  # 256
+
+
+# TODO sagi, maybe set 'clinical' to 0 ?
+if TRAIN_COMMON_PARAMS["model.bb.num_features_clinical"] is None:
+    TRAIN_COMMON_PARAMS["model.bb.num_features"] = TRAIN_COMMON_PARAMS["model.bb.num_features_imaging"]
+else:
+    TRAIN_COMMON_PARAMS["model.bb.num_features"] = (
+        TRAIN_COMMON_PARAMS["model.bb.num_features_imaging"] + TRAIN_COMMON_PARAMS["model.bb.num_features_clinical"]
     )
 
-    dataset_all = Duke.dataset(**params)
-    folds = dataset_balanced_division_to_folds(
-        dataset=dataset_all,
-        output_split_filename=paths["data_split_filename"],
-        keys_to_balance=["data.ground_truth"],
-        workers=0,  # todo: stuck in Export to dataframe
-        nfolds=train_params["data.num_folds"],
-        verbose=True,
-    )
+######################################
+# Inference Common Params
+######################################
+INFER_COMMON_PARAMS = {}
+INFER_COMMON_PARAMS["infer_filename"] = "infer_file.gz"
+INFER_COMMON_PARAMS["checkpoint"] = "best_epoch.ckpt"  # Fuse TIP: possible values are 'best', 'last' or epoch_index.
+INFER_COMMON_PARAMS["data.infer_folds"] = [heldout_fold]  # infer validation set
+INFER_COMMON_PARAMS["data.batch_size"] = 4
+INFER_COMMON_PARAMS["data.num_workers"] = num_workers
+INFER_COMMON_PARAMS["model.classification_task"] = TRAIN_COMMON_PARAMS["model.classification_task"]
+INFER_COMMON_PARAMS["trainer.accelerator"] = "gpu"
+INFER_COMMON_PARAMS["trainer.ckpt_path"] = None
+INFER_COMMON_PARAMS["trainer.num_devices"] = num_devices
 
-    train_sample_ids = []
-    for fold in train_params["data.train_folds"]:
-        train_sample_ids += folds[fold]
-    validation_sample_ids = []
-    for fold in train_params["data.validation_folds"]:
-        validation_sample_ids += folds[fold]
+INFER_COMMON_PARAMS["model.bb.input_channels_num"] = TRAIN_COMMON_PARAMS["model.bb.input_channels_num"]
+INFER_COMMON_PARAMS["model.bb.num_features"] = TRAIN_COMMON_PARAMS["model.bb.num_features"]
+INFER_COMMON_PARAMS["model.post_concat_inputs"] = TRAIN_COMMON_PARAMS["model.post_concat_inputs"]
+INFER_COMMON_PARAMS["model.post_concat_model"] = TRAIN_COMMON_PARAMS["model.post_concat_model"]
+INFER_COMMON_PARAMS["model.dropout_rate"] = TRAIN_COMMON_PARAMS["model.dropout_rate"]
 
-    params["sample_ids"] = train_sample_ids
-    params["reset_cache"] = False
-    params["train"] = True
-    params["cache_kwargs"] = dict(use_pipeline_hash=False, audit_first_sample=False, audit_rate=None)
-    train_dataset = Duke.dataset(**params)
-    # for _ in train_dataset:
-    #     pass
-    params["sample_ids"] = validation_sample_ids
-    params["train"] = False
-    validation_dataset = Duke.dataset(**params)
+######################################
+# Analyze Common Params
+######################################
+EVAL_COMMON_PARAMS = {}
+EVAL_COMMON_PARAMS["infer_filename"] = INFER_COMMON_PARAMS["infer_filename"]
 
-    lgr.info("- Create sampler:")
-    sampler = BatchSamplerDefault(
-        dataset=train_dataset,
-        balanced_class_name="data.ground_truth",
-        num_balanced_classes=train_params["class_num"],
-        batch_size=train_params["data.batch_size"],
-        workers=0,  # train_params['data.train_num_workers'] #todo: stuck
-    )
-    lgr.info("- Create sampler: Done")
 
-    # Create dataloader
-    train_dataloader = DataLoader(
-        dataset=train_dataset,
-        batch_sampler=sampler,
-        collate_fn=CollateDefault(),
-        num_workers=train_params["data.train_num_workers"],
-    )
-    lgr.info("Train Data: Done", {"attrs": "bold"})
-
-    # dataloader
-    validation_dataloader = DataLoader(
-        dataset=validation_dataset,
-        batch_size=train_params["data.batch_size"],
-        collate_fn=CollateDefault(),
-        num_workers=train_params["data.validation_num_workers"],
-    )
-    lgr.info("Validation Data: Done", {"attrs": "bold"})
-
-    # ==============================================================================
-    # Model
-    # ==============================================================================
-    lgr.info("Model:", {"attrs": "bold"})
-
-    conv_inputs = (("data.input.patch_volume", 1),)
+def create_model(
+    conv_inputs: Any,
+    backbone_ch_num: Any,
+    num_backbone_features: Any,
+    post_concat_inputs: Any,
+    post_concat_model: Any,
+    dropout_rate: float,
+) -> torch.nn.Module:
+    """
+    TODO: docu, fix type Any
+    """
     model = Fuse_model_3d_multichannel(
         conv_inputs=conv_inputs,  # previously 'data.input'. could be either 'data.input.patch_volume' or  'data.input.patch_volume_orig'
-        backbone=ResNet(conv_inputs=conv_inputs, ch_num=train_params["backbone_model_dict"]["input_channels_num"]),
+        backbone=ResNet(conv_inputs=conv_inputs, ch_num=backbone_ch_num),
         # since backbone resnet contains pooling and fc, the feature output is 1D,
         # hence we use Head1dClassifier as classification head
         heads=[
             Head1DClassifier(
                 head_name="classification",
-                conv_inputs=[("model.backbone_features", train_params["num_backbone_features"])],
-                post_concat_inputs=train_params["post_concat_inputs"],
-                post_concat_model=train_params["post_concat_model"],
-                dropout_rate=train_params["imaging_dropout"],
+                conv_inputs=[("model.backbone_features", num_backbone_features)],
+                post_concat_inputs=post_concat_inputs,
+                post_concat_model=post_concat_model,
+                dropout_rate=dropout_rate,
                 # append_dropout_rate=train_params['clinical_dropout'],
                 # fused_dropout_rate=train_params['fused_dropout'],
                 shared_classifier_head=None,
@@ -379,135 +221,310 @@ def run_train(
         ],
     )
 
-    lgr.info("Model: Done", {"attrs": "bold"})
+    return model
 
-    # ====================================================================================
-    #  Loss
-    # ====================================================================================
+
+#################################
+# Train Template
+#################################
+def run_train(paths: dict, train_common_params: dict) -> None:
+    # ==============================================================================
+    # Logger
+    # ==============================================================================
+    fuse_logger_start(output_path=paths["model_dir"], console_verbose_level=logging.INFO)
+
+    print("Fuse Train")
+    print(f'model_dir={paths["model_dir"]}')
+    print(f'cache_dir={paths["cache_dir"]}')
+
+    # ==============================================================================
+    # Data
+    # ==============================================================================
+
+    # Holds all datasets' params in one dictionary.
+    # because most of the datasets params are common
+    # TODO: consider use this method. only when runner is running.
+    # For now we'll define it seperatly
+    common_dataset_params = {}
+    # example
+    train_dataset_params = common_dataset_params
+    train_dataset_params["train"] = True
+    validation_dataset_params = common_dataset_params
+    validation_dataset_params["train"] = False
+
+    #### Split Data
+    all_dataset = Duke.dataset(
+        label_type=train_common_params["model.classification_task"],
+        data_dir=paths["data_dir"],
+        cache_dir=paths["cache_dir"],
+        # reset_cache=train_common_params["data.reset_cache"],
+        reset_cache=False,
+        sample_ids=train_common_params["data.sample_ids"],
+        num_workers=train_common_params["data.train_num_workers"],
+        train=False,
+        verbose=False,
+    )
+    folds = dataset_balanced_division_to_folds(
+        dataset=all_dataset,
+        output_split_filename=paths["data_split_filename"],
+        keys_to_balance=["data.ground_truth"],
+        workers=0,  # todo: stuck in Export to dataframe
+        nfolds=train_common_params["data.num_folds"],
+        verbose=True,
+    )
+
+    ## sample ids:
+    train_sample_ids = []
+    for fold in train_common_params["data.train_folds"]:
+        train_sample_ids += folds[fold]
+    validation_sample_ids = []
+    for fold in train_common_params["data.validation_folds"]:
+        validation_sample_ids += folds[fold]
+
+    #### Train Data
+    print("Train Data:")
+
+    ## Create dataset
+    print("- Create Dataset:")
+    train_dataset = Duke.dataset(
+        label_type=train_common_params["model.classification_task"],
+        data_dir=paths["data_dir"],
+        cache_dir=paths["cache_dir"],
+        # reset_cache=train_common_params["data.reset_cache"],
+        reset_cache=False,
+        sample_ids=train_sample_ids,
+        num_workers=train_common_params["data.train_num_workers"],
+        train=True,
+        verbose=False,
+    )
+
+    ## Create batch sampler
+    print("- Create sampler:")
+    sampler = BatchSamplerDefault(
+        dataset=train_dataset,
+        balanced_class_name="data.ground_truth",
+        num_balanced_classes=train_common_params["model.class_num"],
+        batch_size=train_common_params["data.batch_size"],
+        balanced_class_weights=None,
+        # workers = 0 # from michal, not sure if applicable / why
+    )
+
+    print("- Create sampler: Done")
+
+    ## Create dataloader
+    train_dataloader = DataLoader(
+        dataset=train_dataset,
+        shuffle=False,
+        drop_last=False,
+        batch_sampler=sampler,
+        collate_fn=CollateDefault(),
+        num_workers=train_common_params["data.train_num_workers"],
+    )
+    print("Train Data: Done")
+
+    #### Validation data
+    print("Validation Data:")
+
+    validation_dataset = Duke.dataset(
+        label_type=train_common_params["model.classification_task"],
+        data_dir=paths["data_dir"],
+        cache_dir=paths["cache_dir"],
+        # reset_cache=train_common_params["data.reset_cache"],
+        reset_cache=False,
+        sample_ids=validation_sample_ids,
+        num_workers=train_common_params["data.validation_num_workers"],
+        train=False,
+        verbose=False,
+    )
+
+    ## Create dataloader
+    validation_dataloader = DataLoader(
+        dataset=validation_dataset,
+        shuffle=False,
+        drop_last=False,
+        batch_sampler=None,
+        batch_size=train_common_params["data.batch_size"],
+        num_workers=train_common_params["data.validation_num_workers"],
+        collate_fn=CollateDefault(),
+    )
+    print("Validation Data: Done")
+
+    ## Create model
+    print("Model:")
+    model = create_model(
+        conv_inputs=(("data.input.patch_volume", 1),),
+        backbone_ch_num=train_common_params["model.bb.input_channels_num"],
+        num_backbone_features=train_common_params["model.bb.num_features"],
+        post_concat_inputs=train_common_params["model.post_concat_inputs"],
+        post_concat_model=train_common_params["model.post_concat_model"],
+        dropout_rate=train_common_params["model.dropout_rate"],
+    )
+    print("Model: Done")
+
+    # ==========================================================================================================================================
+    #   Loss
+    # ==========================================================================================================================================
     losses = {
         "cls_loss": LossDefault(
             pred="model.logits.classification", target="data.ground_truth", callable=F.cross_entropy, weight=1.0
         ),
     }
 
-    # ====================================================================================
+    # =========================================================================================================
     # Metrics
-    # ====================================================================================
-    lgr.info("Metrics:", {"attrs": "bold"})
-    metrics = OrderedDict([("auc", MetricAUCROC(pred="model.output.classification", target="data.ground_truth"))])
+    # =========================================================================================================
+    train_metrics = OrderedDict([("auc", MetricAUCROC(pred="model.output.classification", target="data.ground_truth"))])
+    validation_metrics = copy.deepcopy(train_metrics)  # use the same metrics in validation as well
+
+    best_epoch_source = dict(monitor="validation.metrics.auc", mode="max")
 
     # =====================================================================================
-    #  Callbacks
+    #  Train - using PyTorch Lightning
+    #  Create training objects, PL module and PL trainer.
     # =====================================================================================
-    callbacks = [
-        # default callbacks
-        # TensorboardCallback(model_dir=paths["model_dir"]),  # save statistics for tensorboard
-        # MetricStatisticsCallback(output_path=paths["model_dir"] + "/metrics.csv"),  # save statistics a csv file
-        # TimeStatisticsCallback(
-        #     num_epochs=train_params["manager.train_params"]["num_epochs"], load_expected_part=0.1
-        # ),  # time profiler
-    ]
-
-    # =====================================================================================
-    #  Manager - Train
-    # =====================================================================================
-    lgr.info("Train:", {"attrs": "bold"})
+    print("Fuse Train:")
 
     # create optimizer
     optimizer = optim.Adam(
-        model.parameters(), lr=train_params["manager.learning_rate"], weight_decay=train_params["manager.weight_decay"]
+        model.parameters(),
+        lr=train_common_params["opt.lr"],
+        weight_decay=train_common_params["opt.weight_decay"],
     )
 
-    # create learning scheduler
-    scheduler = optim.lr_scheduler.ReduceLROnPlateau(optimizer, patience=3, verbose=True)
+    # create scheduler
+    lr_scheduler = optim.lr_scheduler.ReduceLROnPlateau(optimizer)
+    lr_sch_config = dict(scheduler=lr_scheduler, monitor="validation.losses.total_loss")
 
-    # train from scratch
-    manager = "Dummy string for passing flake8"
-    # Providing the objects required for the training process.
-    manager.set_objects(
-        net=model,
-        optimizer=optimizer,
+    # optimizier and lr sch - see pl.LightningModule.configure_optimizers return value for all options
+    optimizers_and_lr_schs = dict(optimizer=optimizer, lr_scheduler=lr_sch_config)
+
+    # create instance of PL module - FuseMedML generic version
+    pl_module = LightningModuleDefault(
+        model_dir=paths["model_dir"],
+        model=model,
         losses=losses,
-        metrics=metrics,
-        best_epoch_source=train_params["manager.best_epoch_source"],
-        lr_scheduler=scheduler,
-        callbacks=callbacks,
-        train_params=train_params["manager.train_params"],
+        train_metrics=train_metrics,
+        validation_metrics=validation_metrics,
+        best_epoch_source=best_epoch_source,
+        optimizers_and_lr_schs=optimizers_and_lr_schs,
     )
 
-    ## Continue training
-    if train_params["manager.resume_checkpoint_filename"] is not None:
-        # Loading the checkpoint including model weights, learning rate, and epoch_index.
-        manager.load_checkpoint(checkpoint=train_params["manager.resume_checkpoint_filename"], mode="train")
+    # create lightining trainer.
+    pl_trainer = pl.Trainer(
+        default_root_dir=paths["model_dir"],
+        max_epochs=train_common_params["trainer.num_epochs"],
+        accelerator=train_common_params["trainer.accelerator"],
+        devices=train_common_params["trainer.num_devices"],
+        auto_select_gpus=True,
+    )
 
-    # Start training
-    manager.train(train_dataloader=train_dataloader, validation_dataloader=validation_dataloader)
+    # train
+    pl_trainer.fit(
+        pl_module, train_dataloader, validation_dataloader, ckpt_path=train_common_params["trainer.ckpt_path"]
+    )
 
-    lgr.info("Train: Done", {"attrs": "bold"})
+    print("Fuse Train: Done")
 
+
+######################################
+# Inference Common Params
+######################################
+# INFER_COMMON_PARAMS = {}
+# INFER_COMMON_PARAMS["data.num_workers"] = TRAIN_COMMON_PARAMS["data.train_num_workers"]
+# INFER_COMMON_PARAMS["data.batch_size"] = 4
+# INFER_COMMON_PARAMS["infer_filename"] = os.path.join(PATHS["inference_dir"], "validation_set_infer.pickle")
+# INFER_COMMON_PARAMS["checkpoint"] = "best"  # Fuse TIP: possible values are 'best', 'last' or epoch_index.
 
 ######################################
 # Inference Template
 ######################################
-def run_infer(paths: dict, infer_common_params: dict, audit_cache: Optional[bool] = True) -> None:
-    #### Logger
+
+
+def run_infer(paths: dict, infer_common_params: dict) -> None:
+    create_dir(paths["inference_dir"])
+    infer_file = os.path.join(paths["inference_dir"], infer_common_params["infer_filename"])
+    checkpoint_file = os.path.join(paths["model_dir"], infer_common_params["checkpoint"])
+
+    ## Logger
     fuse_logger_start(output_path=paths["inference_dir"], console_verbose_level=logging.INFO)
-    lgr = logging.getLogger("Fuse")
-    lgr.info("Fuse Inference", {"attrs": ["bold", "underline"]})
-    lgr.info(
-        f'infer_filename={os.path.join(paths["inference_dir"], infer_common_params["infer_filename"])}',
-        {"color": "magenta"},
-    )
-    lgr.info(f'infer folds={infer_common_params["data.infer_folds"]}', {"color": "magenta"})
+    print("Fuse Inference")
+    print(f"infer_filename={infer_file}")
 
-    ## Data
+    #### create infer dataset
     folds = load_pickle(paths["data_split_filename"])  # assume exists and created in train func
-
     infer_sample_ids = []
     for fold in infer_common_params["data.infer_folds"]:
         infer_sample_ids += folds[fold]
 
-    params = dict(
-        label_type=infer_common_params["classification_task"],
+    infer_dataset = Duke.dataset(
+        label_type=infer_common_params["model.classification_task"],
         data_dir=paths["data_dir"],
         cache_dir=paths["cache_dir"],
-        train=False,
+        # reset_cache=train_common_params["data.reset_cache"],
+        reset_cache=False,
         sample_ids=infer_sample_ids,
+        num_workers=infer_common_params["data.num_workers"],
+        train=False,
         verbose=False,
     )
-    if not audit_cache:
-        params["cache_kwargs"] = dict(use_pipeline_hash=False, audit_first_sample=False, audit_rate=None)
-    else:
-        params["cache_kwargs"] = dict(use_pipeline_hash=False)
-    validation_dataset = Duke.dataset(**params)
 
-    # dataloader
-    validation_dataloader = DataLoader(
-        dataset=validation_dataset,
+    ## Create dataloader
+    infer_dataloader = DataLoader(
+        dataset=infer_dataset,
+        shuffle=False,
+        drop_last=False,
+        batch_sampler=None,
         batch_size=infer_common_params["data.batch_size"],
-        collate_fn=CollateDefault(),
         num_workers=infer_common_params["data.num_workers"],
+        collate_fn=CollateDefault(),
     )
 
-    ## Manager for inference
-    manager = "Dummy string for passing flake8"
-    output_columns = ["model.output.classification", "data.ground_truth"]
-    manager.infer(
-        data_loader=validation_dataloader,
-        input_model_dir=paths["model_dir"],
-        checkpoint=infer_common_params["checkpoint"],
-        output_columns=output_columns,
-        output_file_name=os.path.join(paths["inference_dir"], infer_common_params["infer_filename"]),
+    print("- Create Model:")
+    model = create_model(
+        conv_inputs=(("data.input.patch_volume", 1),),
+        backbone_ch_num=infer_common_params["model.bb.input_channels_num"],
+        num_backbone_features=infer_common_params["model.bb.num_features"],
+        post_concat_inputs=infer_common_params["model.post_concat_inputs"],
+        post_concat_model=infer_common_params["model.post_concat_model"],
+        dropout_rate=infer_common_params["model.dropout_rate"],
     )
+
+    # load python lightning module
+    pl_module = LightningModuleDefault.load_from_checkpoint(
+        checkpoint_file, model_dir=paths["model_dir"], model=model, map_location="cpu", strict=True
+    )
+
+    # set the prediction keys to extract and dump into file (the ones used be the evaluation function).
+    pl_module.set_predictions_keys(["model.output.classification", "data.ground_truth"])
+
+    # create a trainer instance and predict
+    pl_trainer = pl.Trainer(
+        default_root_dir=paths["model_dir"],
+        accelerator=infer_common_params["trainer.accelerator"],
+        devices=infer_common_params["trainer.num_devices"],
+        auto_select_gpus=True,
+    )
+    predictions = pl_trainer.predict(pl_module, infer_dataloader, return_predictions=True)
+
+    # convert list of batch outputs into a dataframe
+    infer_df = convert_predictions_to_dataframe(predictions)
+    save_dataframe(infer_df, infer_file)
+
+    print("Fuse Inference: Done")
 
 
 ######################################
 # Eval Template
 ######################################
-def run_eval(paths: dict, eval_common_params: dict) -> NDict:
+# EVAL_COMMON_PARAMS = {}
+# EVAL_COMMON_PARAMS["infer_filename"] = INFER_COMMON_PARAMS["infer_filename"]
+
+
+def run_eval(paths: dict, eval_common_params: dict) -> None:
     fuse_logger_start(output_path=None, console_verbose_level=logging.INFO)
-    lgr = logging.getLogger("Fuse")
-    lgr.info("Fuse Analyze", {"attrs": ["bold", "underline"]})
+    print("Fuse Eval")
+
+    infer_file = os.path.join(paths["inference_dir"], eval_common_params["infer_filename"])
 
     # metrics
     metrics = OrderedDict(
@@ -532,11 +549,12 @@ def run_eval(paths: dict, eval_common_params: dict) -> NDict:
     # run
     results = evaluator.eval(
         ids=None,
-        data=os.path.join(paths["inference_dir"], eval_common_params["infer_filename"]),
+        data=infer_file,
         metrics=metrics,
         output_dir=paths["eval_dir"],
     )
 
+    print("Fuse Eval: Done")
     return results
 
 
@@ -544,4 +562,23 @@ def run_eval(paths: dict, eval_common_params: dict) -> NDict:
 # Run
 ######################################
 if __name__ == "__main__":
-    main()
+    # allocate gpus
+    NUM_GPUS = 1
+
+    # uncomment if you want to use specific gpus instead of automatically looking for free ones
+    force_gpus = None  # [0]
+    GPU.choose_and_enable_multiple_gpus(NUM_GPUS, force_gpus=force_gpus)
+
+    RUNNING_MODES = ["train", "infer", "eval"]  # Options: 'train', 'infer', 'eval'
+
+    # train
+    if "train" in RUNNING_MODES:
+        run_train(paths=PATHS, train_common_params=TRAIN_COMMON_PARAMS)
+
+    # infer
+    if "infer" in RUNNING_MODES:
+        run_infer(paths=PATHS, infer_common_params=INFER_COMMON_PARAMS)
+
+    # eval
+    if "eval" in RUNNING_MODES:
+        run_eval(paths=PATHS, eval_common_params=EVAL_COMMON_PARAMS)
