+++ examples/fuse_examples/imaging/classification/prostate_x/runner_prostate_x_old.py
@@ -17,37 +17,37 @@ Created on June 30, 2021
 
 """
 
-import torch
 import logging
 import os
-import copy
-from typing import OrderedDict, Optional, List, Dict, Any
+from typing import OrderedDict, Optional, List, Any, Dict
 
 import torch.nn.functional as F
 import torch.optim as optim
 from torch.utils.data.dataloader import DataLoader
 
 import fuse.utils.gpu as GPU
-from fuse_examples.fuse_examples_utils import ask_user, get_fuse_examples_user_dir
+from fuse_examples.fuse_examples_utils import ask_user
 from fuse_examples.imaging.utils.backbone_3d_multichannel import Fuse_model_3d_multichannel, ResNet
 from fuse.data.utils.collates import CollateDefault
 from fuse.data.utils.samplers import BatchSamplerDefault
 from fuse.data.utils.split import dataset_balanced_division_to_folds
 from fuse.dl.losses.loss_default import LossDefault
+
+# from fuse.dl.managers.callbacks.callback_metric_statistics import MetricStatisticsCallback
+# from fuse.dl.managers.callbacks.callback_tensorboard import TensorboardCallback
+# from fuse.dl.managers.callbacks.callback_time_statistics import TimeStatisticsCallback
+# from fuse.dl.managers.manager_default import ManagerDefault
 from fuse.dl.models.heads import Head1DClassifier
 from fuse.eval.evaluator import EvaluatorDefault
 from fuse.eval.metrics.classification.metrics_classification_common import MetricAccuracy, MetricAUCROC, MetricROCCurve
 from fuse.eval.metrics.classification.metrics_thresholding_common import MetricApplyThresholds
+from fuse.utils.file_io.file_io import load_pickle
 from fuse.utils.ndict import NDict
 from fuse.utils.rand.seed import Seed
 from fuse.utils.utils_logger import fuse_logger_start
-from fuse.utils.file_io.file_io import create_dir, load_pickle, save_dataframe
-
 from fuseimg.datasets import prostate_x
 
-from fuse.dl.lightning.pl_module import LightningModuleDefault
-from fuse.dl.lightning.pl_funcs import convert_predictions_to_dataframe
-import pytorch_lightning as pl
+from fuse_examples import fuse_examples_utils
 
 
 def main() -> None:
@@ -58,14 +58,12 @@ def main() -> None:
     if mode == "debug":
         NUM_GPUS = 1
     else:
-        NUM_GPUS = 1
+        NUM_GPUS = 2
     # uncomment if you want to use specific gpus instead of automatically looking for free ones
     force_gpus = None  # [0]
     GPU.choose_and_enable_multiple_gpus(NUM_GPUS, force_gpus=force_gpus)
 
-    PATHS, TRAIN_COMMON_PARAMS, INFER_COMMON_PARAMS, EVAL_COMMON_PARAMS = get_setting(
-        mode, num_devices=NUM_GPUS, n_folds=8, heldout_fold=4
-    )
+    PATHS, TRAIN_COMMON_PARAMS, INFER_COMMON_PARAMS, EVAL_COMMON_PARAMS = get_setting(mode, n_folds=8, heldout_fold=4)
     print(PATHS)
 
     RUNNING_MODES = ["train", "infer", "eval"]  # Options: 'train', 'infer', 'eval'
@@ -89,10 +87,10 @@ def main() -> None:
 
 def get_setting(
     mode: str,
-    num_devices: int,
     label_type: str = prostate_x.ProstateXLabelType.ClinSig,
     n_folds: int = 8,
     heldout_fold: int = 7,
+    num_epoch: Optional[int] = None,
 ) -> List[Dict[str, Any]]:
     ###########################################################################################################
     # Fuse
@@ -109,7 +107,7 @@ def get_setting(
         "PROSTATEX_DATA_PATH" in os.environ
     ), "Expecting environment variable PROSTATEX_DATA_PATH to be set. Follow the instruction in example README file to download and set the path to the data"
     data_dir = os.environ["PROSTATEX_DATA_PATH"]
-    ROOT = os.path.join(get_fuse_examples_user_dir(), "prostate_x")
+    ROOT = os.path.join(fuse_examples_utils.get_fuse_examples_user_dir(), "prostate_x")
 
     if mode == "debug":
         data_split_file = os.path.join(ROOT, f"prostate_x_{n_folds}folds_debug.pkl")
@@ -119,18 +117,21 @@ def get_setting(
         model_dir = os.path.join(ROOT, "model_dir_debug")
         num_workers = 0
         batch_size = 2
-        num_epoch = 5
+        if num_epoch is None:
+            num_epoch = 5
     else:
         data_split_file = os.path.join(ROOT, f"prostatex_{n_folds}folds.pkl")
-        cache_dir = os.path.join(ROOT, "cache_dir_pl")
-        model_dir = os.path.join(ROOT, f"model_dir_pl_{heldout_fold}")
+        cache_dir = os.path.join(ROOT, "cache_dir")
+        model_dir = os.path.join(ROOT, f"model_dir_{heldout_fold}")
         selected_sample_ids = None
 
         num_workers = 16
         batch_size = 50
-        num_epoch = 50
+        if num_epoch is None:
+            num_epoch = 50
     PATHS = {
         "model_dir": model_dir,
+        "force_reset_model_dir": True,  # If True will reset model dir automatically - otherwise will prompt 'are you sure' message.
         "cache_dir": cache_dir,
         "data_split_filename": os.path.join(ROOT, data_split_file),
         "data_dir": data_dir,
@@ -142,6 +143,9 @@ def get_setting(
     # Train Common Params
     ##########################################
     TRAIN_COMMON_PARAMS = {}
+    # ============
+    # Model
+    # ============
 
     # ============
     # Data
@@ -156,20 +160,6 @@ def get_setting(
     TRAIN_COMMON_PARAMS["data.num_folds"] = n_folds
     TRAIN_COMMON_PARAMS["data.train_folds"] = train_folds
     TRAIN_COMMON_PARAMS["data.validation_folds"] = [validation_fold]
-
-    # ===============
-    # PL Trainer
-    # ===============
-    TRAIN_COMMON_PARAMS["trainer.num_epochs"] = num_epoch
-    TRAIN_COMMON_PARAMS["trainer.num_devices"] = num_devices
-    TRAIN_COMMON_PARAMS["trainer.accelerator"] = "gpu"
-    TRAIN_COMMON_PARAMS["trainer.ckpt_path"] = None  # path to the checkpoint you wish continue the training from
-
-    # ===============
-    # Optimizer
-    # ===============
-    TRAIN_COMMON_PARAMS["opt.lr"] = 1e-3
-    TRAIN_COMMON_PARAMS["opt.weight_decay"] = 0.005
 
     # ===============
     # Manager - Train
@@ -191,9 +181,9 @@ def get_setting(
     TRAIN_COMMON_PARAMS["manager.dropout"] = 0.5
     TRAIN_COMMON_PARAMS["manager.momentum"] = 0.9
     TRAIN_COMMON_PARAMS["manager.resume_checkpoint_filename"] = None  # if not None, will try to load the checkpoint
-    # TRAIN_COMMON_PARAMS['imaging_dropout'] = 0.25
-    # # TRAIN_COMMON_PARAMS['fused_dropout'] = 0.0
-    # # TRAIN_COMMON_PARAMS['clinical_dropout'] = 0.0
+    TRAIN_COMMON_PARAMS["imaging_dropout"] = 0.25
+    # TRAIN_COMMON_PARAMS['fused_dropout'] = 0.0
+    # TRAIN_COMMON_PARAMS['clinical_dropout'] = 0.0
 
     TRAIN_COMMON_PARAMS["num_backbone_features_imaging"] = 512
 
@@ -211,7 +201,7 @@ def get_setting(
         )
 
     # classification task:
-    # supported tasks are: 'ClinSig'
+    # supported labels are: 'ClinSig'
     TRAIN_COMMON_PARAMS["label_type"] = label_type
     TRAIN_COMMON_PARAMS["class_num"] = label_type.get_num_classes()
 
@@ -220,31 +210,16 @@ def get_setting(
         "input_channels_num": input_channels_num,
     }
 
-    # ============
-    # Model
-    # ============
-
-    TRAIN_COMMON_PARAMS["model"] = dict(
-        imaging_dropout=0.25,
-        # fused_dropout=0.0,
-        # clinical_dropout=0.0,
-        num_backbone_features=TRAIN_COMMON_PARAMS["num_backbone_features"],
-        input_channels_num=5,
-    )
-
     ######################################
     # Inference Common Params
     ######################################
     INFER_COMMON_PARAMS = {}
-    INFER_COMMON_PARAMS["infer_filename"] = "infer_file.gz"
-    INFER_COMMON_PARAMS["checkpoint"] = "best_epoch.ckpt"
+    INFER_COMMON_PARAMS["infer_filename"] = "validation_set_infer.gz"
+    INFER_COMMON_PARAMS["checkpoint"] = "best"  # Fuse TIP: possible values are 'best', 'last' or epoch_index.
     INFER_COMMON_PARAMS["data.infer_folds"] = [heldout_fold]  # infer validation set
     INFER_COMMON_PARAMS["data.batch_size"] = 4
     INFER_COMMON_PARAMS["data.num_workers"] = num_workers
     INFER_COMMON_PARAMS["label_type"] = TRAIN_COMMON_PARAMS["label_type"]
-    INFER_COMMON_PARAMS["model"] = TRAIN_COMMON_PARAMS["model"]
-    INFER_COMMON_PARAMS["trainer.num_devices"] = num_devices
-    INFER_COMMON_PARAMS["trainer.accelerator"] = "gpu"
 
     ######################################
     # Analyze Common Params
@@ -255,42 +230,12 @@ def get_setting(
     return PATHS, TRAIN_COMMON_PARAMS, INFER_COMMON_PARAMS, EVAL_COMMON_PARAMS
 
 
-def create_model(imaging_dropout: float, num_backbone_features: int, input_channels_num: int) -> torch.nn.Module:
-    """
-    creates the model
-    See Head3DClassifier for details about imaging_dropout, clinical_dropout, fused_dropout
-    """
-    conv_inputs = (("data.input.patch_volume", 1),)
-
-    model = Fuse_model_3d_multichannel(
-        conv_inputs=conv_inputs,  # previously 'data.input'. could be either 'data.input.patch_volume' or  'data.input.patch_volume_orig'
-        backbone=ResNet(conv_inputs=conv_inputs, ch_num=input_channels_num),
-        # since backbone resnet contains pooling and fc, the feature output is 1D,
-        # hence we use Head1dClassifier as classification head
-        heads=[
-            Head1DClassifier(
-                head_name="classification",
-                conv_inputs=[("model.backbone_features", num_backbone_features)],
-                post_concat_inputs=None,  # [('data.clinical_features',9),]
-                post_concat_model=None,  # (256,256)
-                dropout_rate=imaging_dropout,
-                # append_dropout_rate=train_params['clinical_dropout'],
-                # fused_dropout_rate=train_params['fused_dropout'],
-                shared_classifier_head=None,
-                layers_description=None,
-                num_classes=2,
-                # append_features=[("data.input.clinical", 8)],
-                # append_layers_description=(256,128),
-            ),
-        ],
-    )
-    return model
-
-
 #################################
 # Train Template
 #################################
-def run_train(paths: dict, train_params: dict) -> None:
+def run_train(
+    paths: dict, train_params: dict, reset_cache: Optional[bool] = None, audit_cache: Optional[bool] = None
+) -> None:
     Seed.set_seed(222, False)
 
     # ==============================================================================
@@ -311,10 +256,12 @@ def run_train(paths: dict, train_params: dict) -> None:
     # Train Data
     lgr.info("Train Data:", {"attrs": "bold"})
 
-    reset_cache = ask_user("Do you want to reset cache?")
+    if reset_cache is None:
+        reset_cache = ask_user("Do you want to reset cache?")
     cache_kwargs = {"use_pipeline_hash": False}
     if not reset_cache:
-        audit_cache = ask_user("Do you want to audit cache?")
+        if audit_cache is None:
+            audit_cache = ask_user("Do you want to audit cache?")
         if not audit_cache:
             cache_kwargs2 = dict(audit_first_sample=False, audit_rate=None)
             cache_kwargs = {**cache_kwargs, **cache_kwargs2}
@@ -396,7 +343,29 @@ def run_train(paths: dict, train_params: dict) -> None:
     # ==============================================================================
     lgr.info("Model:", {"attrs": "bold"})
 
-    model = create_model(**train_params["model"])
+    conv_inputs = (("data.input.patch_volume", 1),)
+    model = Fuse_model_3d_multichannel(
+        conv_inputs=conv_inputs,  # previously 'data.input'. could be either 'data.input.patch_volume' or  'data.input.patch_volume_orig'
+        backbone=ResNet(conv_inputs=conv_inputs, ch_num=train_params["backbone_model_dict"]["input_channels_num"]),
+        # since backbone resnet contains pooling and fc, the feature output is 1D,
+        # hence we use Head1dClassifier as classification head
+        heads=[
+            Head1DClassifier(
+                head_name="classification",
+                conv_inputs=[("model.backbone_features", train_params["num_backbone_features"])],
+                post_concat_inputs=train_params["post_concat_inputs"],
+                post_concat_model=train_params["post_concat_model"],
+                dropout_rate=train_params["imaging_dropout"],
+                # append_dropout_rate=train_params['clinical_dropout'],
+                # fused_dropout_rate=train_params['fused_dropout'],
+                shared_classifier_head=None,
+                layers_description=None,
+                num_classes=2,
+                # append_features=[("data.input.clinical", 8)],
+                # append_layers_description=(256,128),
+            ),
+        ],
+    )
 
     lgr.info("Model: Done", {"attrs": "bold"})
 
@@ -413,58 +382,54 @@ def run_train(paths: dict, train_params: dict) -> None:
     # Metrics
     # ====================================================================================
     lgr.info("Metrics:", {"attrs": "bold"})
-    train_metrics = OrderedDict([("auc", MetricAUCROC(pred="model.output.classification", target="data.ground_truth"))])
-    validation_metrics = copy.deepcopy(train_metrics)  # use the same metrics in validation as well
-
-    # either a dict with arguments to pass to ModelCheckpoint or list dicts for multiple ModelCheckpoint callbacks (to monitor and save checkpoints for more then one metric).
-    best_epoch_source = dict(
-        monitor="validation.metrics.auc",
-        mode="max",
-    )
-
-    # create optimizer
-    optimizer = optim.SGD(
-        model.parameters(),
-        lr=train_params["opt.lr"],
-        weight_decay=train_params["opt.weight_decay"],
-        momentum=0.9,
-        nesterov=True,
-    )
-
-    # create learning scheduler
-    lr_scheduler = optim.lr_scheduler.ReduceLROnPlateau(optimizer)
-    lr_sch_config = dict(scheduler=lr_scheduler, monitor="validation.losses.total_loss")
-
-    # optimizier and lr sch - see pl.LightningModule.configure_optimizers return value for all options
-    optimizers_and_lr_schs = dict(optimizer=optimizer, lr_scheduler=lr_sch_config)
+    metrics = OrderedDict([("auc", MetricAUCROC(pred="model.output.classification", target="data.ground_truth"))])
 
     # =====================================================================================
-    #  Train
+    #  Callbacks
+    # =====================================================================================
+    callbacks = [
+        # default callbacks
+        # TensorboardCallback(model_dir=paths["model_dir"]),  # save statistics for tensorboard
+        # MetricStatisticsCallback(output_path=paths["model_dir"] + "/metrics.csv"),  # save statistics a csv file
+        # TimeStatisticsCallback(
+        #     num_epochs=train_params["manager.train_params"]["num_epochs"], load_expected_part=0.1
+        # ),  # time profiler
+    ]
+
+    # =====================================================================================
+    #  Manager - Train
     # =====================================================================================
     lgr.info("Train:", {"attrs": "bold"})
 
-    # create instance of PL module - FuseMedML generic version
-    pl_module = LightningModuleDefault(
-        model_dir=paths["model_dir"],
-        model=model,
+    # create optimizer
+    optimizer = optim.Adam(
+        model.parameters(), lr=train_params["manager.learning_rate"], weight_decay=train_params["manager.weight_decay"]
+    )
+
+    # create learning scheduler
+    scheduler = optim.lr_scheduler.ReduceLROnPlateau(optimizer, patience=3, verbose=True)
+
+    # train from scratch
+    manager = "Dummy string for passing flake8"
+    # Providing the objects required for the training process.
+    manager.set_objects(
+        net=model,
+        optimizer=optimizer,
         losses=losses,
-        train_metrics=train_metrics,
-        validation_metrics=validation_metrics,
-        best_epoch_source=best_epoch_source,
-        optimizers_and_lr_schs=optimizers_and_lr_schs,
+        metrics=metrics,
+        best_epoch_source=train_params["manager.best_epoch_source"],
+        lr_scheduler=scheduler,
+        callbacks=callbacks,
+        train_params=train_params["manager.train_params"],
     )
 
-    # create lightining trainer.
-    pl_trainer = pl.Trainer(
-        default_root_dir=paths["model_dir"],
-        max_epochs=train_params["trainer.num_epochs"],
-        accelerator=train_params["trainer.accelerator"],
-        devices=train_params["trainer.num_devices"],
-        auto_select_gpus=True,
-    )
+    ## Continue training
+    if train_params["manager.resume_checkpoint_filename"] is not None:
+        # Loading the checkpoint including model weights, learning rate, and epoch_index.
+        manager.load_checkpoint(checkpoint=train_params["manager.resume_checkpoint_filename"], mode="train")
 
-    # train
-    pl_trainer.fit(pl_module, train_dataloader, validation_dataloader, ckpt_path=train_params["trainer.ckpt_path"])
+    # Start training
+    manager.train(train_dataloader=train_dataloader, validation_dataloader=validation_dataloader)
 
     lgr.info("Train: Done", {"attrs": "bold"})
 
@@ -473,15 +438,14 @@ def run_train(paths: dict, train_params: dict) -> None:
 # Inference Template
 ######################################
 def run_infer(paths: dict, infer_common_params: dict, audit_cache: Optional[bool] = True) -> None:
-    create_dir(paths["inference_dir"])
-    infer_file = os.path.join(paths["inference_dir"], infer_common_params["infer_filename"])
-    checkpoint_file = os.path.join(paths["model_dir"], infer_common_params["checkpoint"])
-
     #### Logger
     fuse_logger_start(output_path=paths["inference_dir"], console_verbose_level=logging.INFO)
     lgr = logging.getLogger("Fuse")
     lgr.info("Fuse Inference", {"attrs": ["bold", "underline"]})
-    lgr.info(f"infer_filename={infer_file}", {"color": "magenta"})
+    lgr.info(
+        f'infer_filename={os.path.join(paths["inference_dir"], infer_common_params["infer_filename"])}',
+        {"color": "magenta"},
+    )
     lgr.info(f'infer folds={infer_common_params["data.infer_folds"]}', {"color": "magenta"})
 
     ## Data
@@ -503,46 +467,32 @@ def run_infer(paths: dict, infer_common_params: dict, audit_cache: Optional[bool
         params["cache_kwargs"] = dict(use_pipeline_hash=False, audit_first_sample=False, audit_rate=None)
     else:
         params["cache_kwargs"] = dict(use_pipeline_hash=False)
-    infer_dataset = prostate_x.ProstateX.dataset(**params)
+    validation_dataset = prostate_x.ProstateX.dataset(**params)
 
     # dataloader
-    infer_dataloader = DataLoader(
-        dataset=infer_dataset,
+    validation_dataloader = DataLoader(
+        dataset=validation_dataset,
         batch_size=infer_common_params["data.batch_size"],
         collate_fn=CollateDefault(),
         num_workers=infer_common_params["data.num_workers"],
     )
 
-    # load python lightning module
-    model = create_model(**infer_common_params["model"])
-    pl_module = LightningModuleDefault.load_from_checkpoint(
-        checkpoint_file, model_dir=paths["model_dir"], model=model, map_location="cpu", strict=True
+    ## Manager for inference
+    manager = "Dummy string for passing flake8"
+    output_columns = ["model.output.classification", "data.ground_truth"]
+    manager.infer(
+        data_loader=validation_dataloader,
+        input_model_dir=paths["model_dir"],
+        checkpoint=infer_common_params["checkpoint"],
+        output_columns=output_columns,
+        output_file_name=os.path.join(paths["inference_dir"], infer_common_params["infer_filename"]),
     )
-    # set the prediction keys to extract (the ones used be the evaluation function).
-    pl_module.set_predictions_keys(
-        ["model.output.classification", "data.ground_truth"]
-    )  # which keys to extract and dump into file
-
-    # create a trainer instance
-    pl_trainer = pl.Trainer(
-        default_root_dir=paths["model_dir"],
-        accelerator=infer_common_params["trainer.accelerator"],
-        devices=infer_common_params["trainer.num_devices"],
-        auto_select_gpus=True,
-    )
-    predictions = pl_trainer.predict(pl_module, infer_dataloader, return_predictions=True)
-
-    # convert list of batch outputs into a dataframe
-    infer_df = convert_predictions_to_dataframe(predictions)
-    save_dataframe(infer_df, infer_file)
 
 
 ######################################
 # Analyze Template
 ######################################
 def run_eval(paths: dict, eval_common_params: dict) -> NDict:
-    infer_file = os.path.join(paths["inference_dir"], eval_common_params["infer_filename"])
-
     fuse_logger_start(output_path=None, console_verbose_level=logging.INFO)
     lgr = logging.getLogger("Fuse")
     lgr.info("Fuse Analyze", {"attrs": ["bold", "underline"]})
@@ -568,7 +518,12 @@ def run_eval(paths: dict, eval_common_params: dict) -> NDict:
     evaluator = EvaluatorDefault()
 
     # run
-    results = evaluator.eval(ids=None, data=infer_file, metrics=metrics, output_dir=paths["eval_dir"])
+    results = evaluator.eval(
+        ids=None,
+        data=os.path.join(paths["inference_dir"], eval_common_params["infer_filename"]),
+        metrics=metrics,
+        output_dir=paths["eval_dir"],
+    )
 
     return results
